refactor: completion module, cli test helpers, phase 5 polish
- Extract shell completion generators to src/completion.ts; index imports
  generateCompletionScript
- withCommandContext: generic tuple/return typing (no explicit any)
- docs: CommanderCommandInternals + asCommanderInternals for _hidden/_args
- security: document why INVESTEC_CLIENT_ID is omitted from secret env list;
  test that it is not reported as a secret
- test/helpers/cli-mocks.ts: getRuntimeCredentialsMock, testCredentials,
  createSpinnerControlMock; migrate key tests to async vi.mock factories
- test/utils/completion.test.ts for bash/zsh output and invalid shell
- IMPROVEMENT_PLAN: phase 5 complete

Made-with: Cursor
