+++ IMPROVEMENT_PLAN.md
@@ -85,7 +85,7 @@ Phased plan from the architecture / error-handling review. **Status** is updated
 | **2** | Error model: `fetch.ts` codes, `credentials-store` `CliError`, `determineExitCode` heuristics, `generateCompletionScript` | **Done** — `ERROR_CODES.UNSUPPORTED_OPERATION`; fetch uses `UNSUPPORTED_OPERATION` / `INVESTEC_API_ERROR`; `readCredentialsFile` / `loadCredentialsFile` throw `CliError`; `determineExitCode` uses code map then narrower heuristics (exported); completion uses `CliError` + `determineExitCode` |
 | **3** | Static `utils` imports in commands; single `configureChalk` call site | **Done** — all commands use static `isStdoutPiped` / `readStdin` from `utils.js` (no dynamic `import('../utils.js')`); `configureChalk()` only invoked from `index.ts` at startup |
 | **4** | Tests: `bank`, `register`, `login`, `ai`; optional shared mock harness | **Done** (tests added; shared mock harness still optional) |
-| **5** | Polish: `withCommandContext` generics, `docs.ts` typing, `INVESTEC_CLIENT_ID` policy, split `index.ts` | Pending |
+| **5** | Polish: `withCommandContext` generics, `docs.ts` typing, `INVESTEC_CLIENT_ID` policy, split `index.ts` | **Done** — see below |
 
 ### Phase 1 details (done in tree)
 
@@ -116,6 +116,11 @@ Phased plan from the architecture / error-handling review. **Status** is updated
 - `test/cmds/ai.test.ts` — valid structured JSON → `writeFile`; missing/invalid response → `INVALID_INPUT`.
 - Optional: `test/helpers/cli-mocks.ts` for shared `runtime-credentials` / spinner mocks (not added).
 
-### Phase 5
+### Phase 5 (done)
 
-- Tighten types in `withCommandContext`, `docs.ts`; document or adjust `SECRET_ENV_VARS` for client id; extract completion / registration from `index.ts`.
+- **`src/completion.ts`** — `generateCompletionScript` + bash/zsh generators moved out of `index.ts`.
+- **`withCommandContext`** — generic `A extends readonly unknown[], R` (no `any`); same runtime behavior.
+- **`docs.ts`** — `CommanderCommandInternals` + `asCommanderInternals()` instead of `(command as any)`.
+- **`security.ts`** — JSDoc: why `INVESTEC_CLIENT_ID` is not in `SECRET_ENV_VARS`; test asserts it is not flagged.
+- **`test/helpers/cli-mocks.ts`** — `getRuntimeCredentialsMock`, `testCredentials`, `createSpinnerControlMock`; consumers use **async** `vi.mock(…, async () => import('./cli-mocks.js'))` to satisfy Vitest hoisting.
+- **`test/utils/completion.test.ts`** — smoke tests for completion output + invalid shell.
