+++ test/cmds/register.test.ts
@@ -10,10 +10,10 @@ vi.mock('node-fetch', () => ({
   default: mockFetch,
 }));
 
-vi.mock('../../src/runtime-credentials.ts', () => ({
-  printTitleBox: vi.fn(),
-  optionCredentials: vi.fn(async (_options, credentials) => credentials),
-}));
+vi.mock('../../src/runtime-credentials.ts', async () => {
+  const { getRuntimeCredentialsMock } = await import('../helpers/cli-mocks.js');
+  return getRuntimeCredentialsMock();
+});
 
 vi.mock('../../src/utils.ts', async () => {
   const actual = await vi.importActual<typeof import('../../src/utils.ts')>('../../src/utils.ts');
